# backend/api-gateway/swagger.yaml
openapi: 3.0.0
info:
  title: ForkLine API Gateway
  description: |
    Complete API documentation for the ForkLine restaurant management system.
    This API Gateway routes requests to multiple microservices:
    - Auth Service (Port 3001)
    - User Service (Port 3015)  
    - Restaurant Service (Port 3012)
    - Order Service (Port 3010)
    - Booking Service (Port 3011)
    - And more...
  version: 1.0.0
  contact:
    name: ForkLine API Support
    url: https://github.com/your-repo/forkline
    email: support@forkline.com
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.forkline.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    Error:
      type: object
      properties:
        status:
          type: string
          enum: [fail, error]
          example: "error"
        message:
          type: string
          example: "An error occurred"
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
          example: "success"
        message:
          type: string
          example: "Operation completed successfully"

    # Auth schemas
...
    Restaurant:
      type: object
      properties:
        id:
          type: string
          example: "64f8b9c12345678901234567"
        name:
          type: string
          example: "Mario's Italian Restaurant"
        description:
          type: string
          example: "Authentic Italian cuisine in the heart of the city"
        cuisine_type:
          type: string
          example: "Italian"
        address:
          type: object
          properties:
            street:
              type: string
              example: "123 Main St"
            city:
              type: string
              example: "New York"
            state:
              type: string
              example: "NY"
            zip_code:
              type: string
              example: "10001"
        location:
          type: object
          properties:
            type:
              type: string
              enum: [Point]
              example: "Point"
            coordinates:
              type: array
              items:
                type: number
              example: [-73.935242, 40.730610]
        rating:
          type: number
          format: float
          example: 4.5
        price_range:
          type: string
          enum: ["$", "$$", "$$$", "$$$$"]
          example: "$$"
        has_delivery:
          type: boolean
          example: true
        has_reservations:
          type: boolean
          example: true
        status:
          type: string
          enum: [active, inactive, pending]
          example: "active"
        owner_id:
          type: string
          example: "64f8b9c12345678901234567"
        created_at:
          type: string
          format: date-time
          example: "2023-09-01T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-09-01T10:00:00Z"
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        pagination:
          type: object
          properties:
            current_page:
              type: integer
              example: 1
            per_page:
              type: integer
              example: 10
            total:
              type: integer
              example: 100
            total_pages:
              type: integer
              example: 10
            has_next:
              type: boolean
              example: true
            has_prev:
              type: boolean
              example: false
security:
  - bearerAuth: []
paths:
...
  /api/restaurants:
    get:
      tags:
        - Restaurant
      summary: Get restaurants
      description: Get paginated list of restaurants with optional filtering
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: Items per page
          schema:
            type: integer
            default: 10
        - name: cuisine_type
          in: query
          description: Filter by cuisine type
          schema:
            type: string
            example: "Italian"
        - name: min_rating
          in: query
          description: Minimum rating filter
          schema:
            type: number
            format: float
            example: 4.0
        - name: has_delivery
          in: query
          description: Filter by delivery availability
          schema:
            type: boolean
        - name: search
          in: query
          description: Search restaurants by name or description
          schema:
            type: string
        - name: lat
          in: query
          description: Latitude for location-based search
          schema:
            type: number
            format: float
        - name: lng
          in: query
          description: Longitude for location-based search
          schema:
            type: number
            format: float
        - name: max_distance
          in: query
          description: Maximum distance in kilometers
          schema:
            type: number
            format: float
            default: 10
      responses:
        '200':
          description: Restaurants retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Restaurant'

    post:
      tags:
        - Restaurant
      summary: Create restaurant
      description: Create a new restaurant (requires restaurant owner role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRestaurantRequest'
      responses:
        '201':
          description: Restaurant created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    $ref: '#/components/schemas/Restaurant'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/restaurants/{restaurant_id}:
    get:
      tags:
        - Restaurant
      summary: Get restaurant by ID
      description: Get detailed information about a specific restaurant
      parameters:
        - name: restaurant_id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
      responses:
        '200':
          description: Restaurant retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    $ref: '#/components/schemas/Restaurant'
        '404':
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Restaurant
      summary: Update restaurant
      description: Update restaurant information (requires restaurant owner role)
      parameters:
        - name: restaurant_id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRestaurantRequest'
      responses:
        '200':
          description: Restaurant updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    $ref: '#/components/schemas/Restaurant'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Restaurant
      summary: Delete restaurant
      description: Delete a restaurant (requires restaurant owner role)
      parameters:
        - name: restaurant_id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
      responses:
        '200':
          description: Restaurant deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/restaurants/{restaurant_id}/menu-items:
    get:
      tags:
        - Menu
      summary: Get menu items
      description: Get paginated list of menu items for a restaurant
      parameters:
        - name: restaurant_id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: Items per page
          schema:
            type: integer
            default: 10
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
        - name: is_popular
          in: query
          description: Filter by popular items
          schema:
            type: boolean
        - name: min_price
          in: query
          description: Minimum price filter
          schema:
            type: number
            format: float
        - name: max_price
          in: query
          description: Maximum price filter
          schema:
            type: number
            format: float
        - name: search
          in: query
          description: Search menu items by name
          schema:
            type: string
      responses:
        '200':
          description: Menu items retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/MenuItem'

    post:
      tags:
        - Menu
      summary: Create menu item
      description: Create a new menu item for a restaurant (requires restaurant staff role)
      parameters:
        - name: restaurant_id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
                - price
                - category
              properties:
                name:
                  type: string
                  example: "Margherita Pizza"
                description:
                  type: string
                  example: "Classic pizza with tomato sauce and mozzarella"
                price:
                  type: number
                  format: float
                  example: 12.99
                category:
                  type: string
                  example: "Pizza"
                image_url:
                  type: string
                  example: "https://example.com/pizza.jpg"
                is_spicy:
                  type: boolean
                  example: false
                is_vegetarian:
                  type: boolean
                  example: true
                allergens:
                  type: array
                  items:
                    type: string
                  example: ["gluten", "dairy"]
      responses:
        '201':
          description: Menu item created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    $ref: '#/components/schemas/MenuItem'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/restaurants/{restaurant_id}/tables:
    get:
      tags:
        - Table
      summary: Get tables
      description: Get list of tables for a restaurant
      parameters:
        - name: restaurant_id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by table status
          schema:
            type: string
            enum: [available, occupied, reserved]
        - name: min_capacity
          in: query
          description: Minimum table capacity
          schema:
            type: integer
      responses:
        '200':
          description: Tables retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        table_number:
                          type: string
                        capacity:
                          type: integer
                        status:
                          type: string
                          enum: [available, occupied, reserved]

    post:
      tags:
        - Table
      summary: Create table
      description: Create a new table for a restaurant (requires restaurant staff role)
      parameters:
        - name: restaurant_id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - table_number
                - capacity
              properties:
                table_number:
                  type: string
                  example: "T-01"
                capacity:
                  type: integer
                  example: 4
                table_type:
                  type: string
                  example: "indoor"
      responses:
        '201':
          description: Table created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      table_number:
                        type: string
                      capacity:
                        type: integer
                      status:
                        type: string
...